moved linting result object filter to its own module
